try eliminate warning on windows

+++ include/aris/dynamic/mechanism_puma_5axis.hpp
@@ -19,7 +19,7 @@ namespace aris::dynamic
 			dynamic_cast<GeneralMotion&>(model()->generalMotionPool()[0]).setMpm(ee_pm);
 			if (ext_axes)
 			{
-				for (int i = 6; i < model()->motionPool().size(); ++i)
+				for (int i = 6; i < static_cast<int>(model()->motionPool().size()); ++i)
 				{
 					model()->motionPool()[i].setMp(ext_axes[i - 6]);
 				}
